requirements.txt
# requirements.txt
Django==1.6.6
Markdown==2.4.1
stripe==1.19.0
django-jsonfield==0.9.10
south==1.0
django-suit==0.2.3
django-debug-toolbar==0.9.4
gunicorn
django-codemirror-widget==0.4.0
django-markupfield==1.2.1
Pillow=2.5.3